api/requirements.txt
fastapi
uvicorn
pandas
kafka-python